refactor(tests): lift stub backend into shared tests/harness module
- Move the prompt-dispatching _StubBackend and its install/silence/force_interactive
  helpers out of tests/test_deep_orchestrator.py into the canonical shared
  harness module tests/harness/stub_backend.py (StubBackend, install_stub_backend,
  silence, force_interactive, PARTIAL_FIX_MARKER).
- Re-export the moved symbols under their historical _-private names in
  test_deep_orchestrator.py so the many sibling modules that
  `from tests.test_deep_orchestrator import _StubBackend, _silence, ...`
  keep working unchanged.
- Repoint tests/test_archive_data_capture.py to import the stub directly from
  the harness instead of reaching across test modules.

Daydream-Run: 20260727174106-d36c9421
Daydream-Version: 0.25.0
Entire-Checkpoint: babea412b45c
